Switch from storing state in S3 to using cookies
This changes from using S3 to store state to instead storing state in a cookie.

This also adds new behavior to both inform AWS of the issuer URL (the URL of
this website/relying party) but also to parse the referer when a user is sent
to this RP after their federated session has expired.

This way a user is given a link to click when their session expires which will
re-authenticate them and send them back to the AWS product that they were using
when their session expired

* Make logging log level configurable with an environment variable
* Change the /pick_role argument from "role" to "role_arn" to match the other endpoints