Update UI to match Mozilla AWS CLI
This changes the UI and flow of federated-aws-rp from a syncronous
backend web flow to an asynchronous frontend + backend web flow.

This allows re-using the UI from the Mozilla AWS CLI so that they
match.

A side effect of this will be to fix #5 by logging the user out before
logging them in.

To do this I've made the following changes
* Enabled sending binary data to users to support sending font files
* Added support for POSTing to the RP
* Added API endpoints to mirror the API endpoints in mozilla-aws-cli
* Add support for using account aliases in the "account" query argument
  utils.py:convert_account_alias
* Add sorting and grouping of roles by account for presentation in the
  role picker
* Added support for reading files from the package
* Included static assets (js, html, fonts) copied from mozilla-aws-cli
* Added a fake /api/heartbeat endpoint since we don't need one like maws
  does

This is a bit rewrite of the project and affects a lot of surface area.

Fixes #7