Use AWS account alias instead of ID for issuer URL
This begins passing the AWS account alias to the backend when a role is
picked so that it can be used when constructing the issuer URL. The issuer
URL is the URL the user is shown by AWS after their session has expired.
If the user clicks on the URL it will take them back to federated-aws-rp
to have their session refreshed.

By using the AWS account alias in this URL instead of the AWS account ID
it should be clearer to the user which account it is that they're about
to establish a refreshed session with.