Add account id to account name
I am often confused searching the login page for the AWS Account ID that
corresponds to the Role I want to select. This simply add the Account ID
to the output for faster finding. Open to suggestions or feedback if
folks have other workflows that solve this.

Command Line Version of this request:
https://github.com/mozilla-iam/mozilla-aws-cli/pull/241